requirements.txt
streamlit
PyPDF2
fastapi
uvicorn
python-multipart
google-generativeai